refactor: 펫 도메인 서비스 계층 Command/Query 분리
- 기존 펫 서비스 로직을 읽기(Query)와 쓰기(Command) 책임으로 분리
- PetCommandService: 펫 등록, 수정, 삭제 등 상태를 변경하는 비즈니스 로직 담당
- PetQueryService: 펫 목록, 상세 조회, 몸무게 로그 조회 등 읽기 전용 로직 담당 및 @Transactional(readOnly = true) 적용으로 성능 최적화
- PetController: 분리된 서비스에 맞게 의존성 주입 및 엔드포인트 호출 로직 매핑
